Add MarshalText() to schema structs
When e.g. GroupVersion is used as map key and the map as a whole is
passed to json.Marshal() it fails with:

`internal error: json: unsupported type: map[schema.GroupVersion]error`

The marshaling can happen indirectly via klog.Logger.Info() and friends,
which leads to useless log messages.

This commit adds MarshalText() methods to each struct which simply
return String() as []byte.